Conflicts:
	Library/Formula/qt.rb

+++ Library/Formula/qt.rb
@@ -56,7 +56,7 @@ class Qt < Formula
             "-confirm-license", "-opensource",
             "-cocoa", "-fast" ]
 
-    # we have to disable all tjos to avoid triggering optimization code
+    # we have to disable 3DNow! to avoid triggering optimization code
     # that will fail with clang. Only seems to occur in superenv, perhaps
     # because we rename clang to cc and Qt thinks it can build with special
     # assembler commands. In --env=std, Qt seems aware of this.)
@@ -77,6 +77,8 @@ class Qt < Formula
     #if build.with? 'qtdbus'
       args << "-I#{Formula.factory('d-bus').lib}/dbus-1.0/include"
       args << "-I#{Formula.factory('d-bus').include}/dbus-1.0"
+      args << "-L#{Formula.factory('d-bus').lib}"
+      args << "-ldbus-1"
     # forced end
 
     # forced if build.with? 'qt3support'
@@ -120,14 +122,11 @@ class Qt < Formula
     (bin+'pixeltool.app').rmtree
     (bin+'qhelpconverter.app').rmtree
     # remove porting file for non-humans
-    (prefix+'q3porting.xml').unlink
+    (prefix+'q3porting.xml').unlink if build.without? 'qt3support'
 
     # Some config scripts will only find Qt in a "Frameworks" folder
-    # VirtualBox is an example of where this is needed
-    # See: https://github.com/mxcl/homebrew/issues/issue/745
-    cd prefix do
-      ln_s lib, prefix + "Frameworks"
-    end
+    frameworks.mkpath
+    ln_s Dir['lib/*.framework'], frameworks
 
     # The pkg-config files installed suggest that headers can be found in the
     # `include` directory. Make this so by creating symlinks from `include` to
